Updated the context changes and ux changes
