+++ template/src/subComponents/SidePanelHeader.tsx
@@ -111,8 +111,8 @@ export const SidePanelStyles = StyleSheet.create({
     flexDirection: 'row',
     justifyContent: 'space-between',
     alignItems: 'center',
-    paddingHorizontal: $config.ENABLE_CONVERSATIONAL_AI ? 20 : 16,
-    paddingVertical: $config.ENABLE_CONVERSATIONAL_AI ? 20 : 22,
+    paddingHorizontal: 20,
+    paddingVertical: 12,
     borderBottomWidth: 1,
     borderBottomColor: $config.CARD_LAYER_3_COLOR,
     position: 'relative',
